refactor: step3 완료, 리사이클러뷰 리팩토링
