+++ app/src/main/java/woowacourse/movie/view/movies/cinema/CinemaViewHolder.kt
@@ -1,0 +1,18 @@
+package woowacourse.movie.view.movies.cinema
+
+import androidx.recyclerview.widget.RecyclerView.ViewHolder
+import woowacourse.movie.databinding.ItemCinemaBinding
+import woowacourse.movie.domain.model.Screening
+
+class CinemaViewHolder(
+    val binding: ItemCinemaBinding,
+) : ViewHolder(binding.root) {
+    fun bind(
+        screening: Screening,
+        eventListener: OnCinemaSelectionListener,
+    ) {
+        binding.screening = screening
+        binding.eventListener = eventListener
+        binding.executePendingBindings()
+    }
+}
